Convert ArbitraryPhantomWrapper

+++ src/ArbitraryPhantomWrapper.ts
@@ -20,24 +20,15 @@ export {
  * Wraps an arbitrary object with a PhantomCore instance.
  */
 export default class ArbitraryPhantomWrapper extends PhantomCore {
-  /**
-   * @param {any} wrappedValue
-   * @param {Object} options? [default = {}]
-   */
-  constructor(wrappedValue, options = {}) {
+  protected _wrappedValue: unknown = null;
+
+  constructor(wrappedValue: unknown, options = {}) {
     super(options);
 
-    this._wrappedValue = null;
     this._setWrappedValue(wrappedValue);
   }
 
-  /**
-   * Internally sets the wrapped element.
-   *
-   * @param {any} wrappedValue
-   * @return {void}
-   */
-  _setWrappedValue(wrappedValue) {
+  _setWrappedValue(wrappedValue: unknown) {
     if (this._wrappedValue) {
       throw new Error("_setWrappedValue cannot be called more than once");
     }
@@ -64,9 +55,11 @@ export default class ArbitraryPhantomWrapper extends PhantomCore {
    * prior to normal destruct operations for this class.
    * @return {Promise<void>}
    */
-  async destroy(destroyHandler = () => null) {
+  async destroy(destroyHandler?: () => void) {
     return super.destroy(async () => {
-      await destroyHandler();
+      if (typeof destroyHandler === "function") {
+        await destroyHandler();
+      }
 
       // IMPORTANT: Setting this AFTER super destroy so that it can potentially
       // be intercepted by other destruct handlers in extension classes
